To keep our Childs safe online, there is no allternative of Parent's Involvements.
We are leaving in the Digital World, surounded by digital information
24/7. We can’t and should not keep our child away from this Digital
World. If we do, our child will be left behind. Education is also in
the Digital format. In US the school has moved to digital format. In
near future, our crounty will be the same. There are a lot of
advantages in Learning with the help of Technology. But at the same
time there are some bad things too. Our responsiblity is to teach our
kids to be a “Good Digital Citizen” just like we teach to be a Good
Citizen.